FROMLIST: btusb: wake system up when receives a wake irq
Currrently we are disabling this wake irq after receiving it. If this
happens before we finish suspend and the pm event check is disabled,
the system will continue suspending, and this irq would not work again.

We may need to abort system suspend to avoid that.

Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
(am from https://patchwork.kernel.org/patch/9589455/)

BUG=b:35587187
TEST=Build and boot

Change-Id: I3bdbf857c422404668366c5a8cd6855d380d37f7
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/446200
Reviewed-by: Brian Norris <briannorris@chromium.org>