UPSTREAM: Bluetooth: btusb: wake system up when receives a wake irq
Currrently we are disabling this wake irq after receiving it. If this
happens before we finish suspend and the pm event check is disabled,
the system will continue suspending, and this irq would not work again.

We may need to abort system suspend to avoid that.

BUG=chromium:707244
TEST=Build and run

Change-Id: I13ce8142e431bedc3514130cc01747aa09946c97
Signed-off-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 017789f37b9b)