cpu-boost: Rework scheduling setup
This patch minimizes the latency of an input
boost by creating a dedicated kworker and scheduling
it with the realtime policy "SCHED_FIFO".
It is known that workqueues can be delayed
by a significant amount of time.

We use "MAX_RT_PRIO - 2" to have the input
boost task preempt userspace RT tasks
if needed while being preempted by kernel
internal RT tasks which are scheduled at
"MAX_RT_PRIO - 1" or higher.

Also since now the cpu-boost wq would
only handle the work to disable a running
input boost, which isn't latency critical,
we break up the cpu-boost wq and move the
work to disable a running input boost
to the shared regular wq.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>