msm: ipa: Fix handling for failed memory allocation
If kzalloc fails for cmd2 in this code path, we
end up assuming that cmd2 does not equal NULL.
The check after the memory allocation for cmd2 looks
for cmd1 which is wrong because cmd1 cannot equal
NULL at that point.

This patch corrects the check to fix the handling
for failed memory allocation in this code path.

Fixes: cb8ff96f8fa2 ("msm: ipa: Remove usage of stack memory")
Change-Id: I94799d82ecc67523db33756896cc7774ff80a456
Signed-off-by: Alex Naidis <alex.naidis@linux.com>